Modified paulp's template project to be an SBT plugin
